fix: Rename server files
